refactor: split OocMuteCommands.cs into 3 files
